2024-03-16 중간 작업 커밋

+++ src/main/resources/sql/create_t_accounting_types.sql
@@ -10,7 +10,7 @@ CREATE SEQUENCE public.t_accounting_types_seq
 CREATE TABLE public.t_accounting_types (
 	act_idx int8 DEFAULT nextval('t_accounting_types_seq'::regclass) NOT NULL,
 	user_idx int8 NOT NULL,
-	act_parent_idx int4 NULL,
+	act_parent_idx int8 NULL,
 	act_either varchar(1) NULL,
 	act_ctg_name varchar(50) NULL,
 	act_seq int4 NULL,
